requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
django-allauth==65.14.3
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
